Change in symatics for server wrapper
The server wrapper will now manage both the server process as well
as the plugin process and write to each others stdin manually vs
using a unix socket.  The server process is not started by default,
only the plugin process is which will (in the future) then inturn
start the server process after reading the relevant configuration
files.  The plugin functiona and the server wrapper function need
to be split into separate files so that they do not require each
other when being run.  This will make the server wrapper process
very lightweight (and not need to be restarted often)